# docs/swagger.yaml
definitions:
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.MessageResponse:
    properties:
      message:
        type: string
    type: object
  handlers.PermissionResponse:
    properties:
      access_type:
        type: string
      document_id:
        type: string
      id:
        type: string
      user_email:
        type: string
      user_id:
        type: string
    type: object
...
  handlers.RemoveAccessRequest:
    properties:
      user_email:
        type: string
    required:
    - user_email
    type: object
  handlers.ShareDocumentRequest:
    properties:
      access_type:
        type: string
      user_email:
        type: string
    required:
    - access_type
    - user_email
    type: object
...
info:
  contact: {}
paths:
...
  /documents/{id}/permissions:
    get:
      consumes:
      - application/json
      description: Get all permissions for a specific document
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.PermissionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get permissions of a document
      tags:
      - Permissions
  /documents/{id}/permissions/remove:
    post:
      consumes:
      - application/json
      description: Remove a specific user's access to a document
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      - description: Remove access request
        in: body
        name: RemoveAccessRequest
        required: true
        schema:
          $ref: '#/definitions/handlers.RemoveAccessRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Remove a user's access to a document
      tags:
      - Permissions
  /documents/{id}/permissions/share:
    post:
      consumes:
      - application/json
      description: Share a document with a user by providing the access type (read,
        edit, admin)
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      - description: Share document request
        in: body
        name: ShareDocumentRequest
        required: true
        schema:
          $ref: '#/definitions/handlers.ShareDocumentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.PermissionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Share a document with a user
      tags:
      - Permissions
swagger: "2.0"